requirements.txt
keras==2.10.0
keras==2.15.0
numpy==1.19.2
Pillow==10.0.1
Pillow==9.5.0
Pillow==10.2.0
streamlit==1.30.0
